src/main/java/cart/domain/Item.java
package cart.domain;

public class Item {
    private final Product product;
    private int quantity;

    public Item(Product product) {
        this(product, 1);
    }

    public Item(Product product, int quantity) {
        this.product = product;
        this.quantity = quantity;
    }

    public void changeQuantity(int quantity) {
        this.quantity = quantity;
    }

    public Money calculateItemPrice() {
        return product.getPrice().multiply(quantity);
    }

    public Product getProduct() {
        return product;
    }

    public int getQuantity() {
        return quantity;
    }
}
